backend/requirements.txt
alabaster==0.7.12
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
attrs==21.2.0
Babel==2.8.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.3
Brlapi==0.8.3
Brotli==1.0.9
cachetools==4.2.4
canvasapi==3.3.0
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.6
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.4.2
docutils==0.17.1
duplicity==0.8.21
exceptiongroup==1.1.3
fastapi==0.104.1
fasteners==0.14.1
Flask==3.0.3
fonttools==4.50.0
future==0.18.2
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.61.0
gpg==1.16.0
grpcio==1.59.2
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.2
idna==3.4
imagekitio==3.2.0
imagesize==1.3.0
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jiter==0.6.1
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
lockfile==0.12.2
lorem==0.1.1
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib==3.8.3
meson==0.61.2
monotonic==1.6
more-itertools==8.10.0
mutagen==1.45.1
netifaces==0.11.0
nftables==0.1
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openai==1.52.0
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.5.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0.1
reportlab==3.6.8
requests==2.31.0
requests-toolbelt==0.10.1
roman==3.3
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==4.3.2
starlette==0.27.0
systemd-python==234
termcolor==2.4.0
tqdm==4.66.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==1.26.18
usb-creator==0.3.7
uvicorn==0.24.0
vboxapi==1.0
wadllib==1.3.6
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.0
WoeUSB-ng==0.2.12
wxPython==4.0.7
xdg==5
xkit==0.0.0
yt-dlp==2022.4.8
zipp==1.0.0
